refactor(reservation): 라이프사이클 트랜잭션 범위를 외부 API 호출 이후로 분리
